Allows marking media as shared and/or for direct messages
Media uploads support a `media_category` value, which may be one of:

- `tweet_image`
- `tweet_gif`
- `tweet_video`
- `dm_image`
- `dm_gif`
- `dm_video`

Additionally, when uploaded for a direct message, they may be marked as
`shared`. This patch updates the media classes to allow passing along
these parameters. The `media_category` is detected based on the provided
`$mediaType`, defaulting to an image; the prefix is based on whether or
not it is marked for usage with a direct message. The `shared` flag is
only considered when the media is marked to use with a DM.