Adds support for uploaded media attachments to direct messages
Adds a `directMessagesEventsNew()` method, which uses the
`direct_messages/events/new` endpoint  in order to send a direct message
to a user. The `directMessagesNew()` method has been updated to proxy to
the new method, as the `direct_messages/new` endpoint is deprecated.
This change was made both due to deprecation of the old functionality,
and to provide support for media attachments in direct messages.

This method _requires_ a recipient **identifier**, not a screen name. As
such, the method now checks for an identifier passed to `$user`; if it
is not, it uses `usersShow()` to look it up before sending the DM.

The direct message uses the new JSON structure for the
`direct_messages/events/new` endpoint for sending the message text. If
`$extraParams` includes a `media_id`, this is then provided as a message
attachment.